drivers/hv/mshv_vtl: detect dropped sidecar remove signal
In automated testing, it looks like perhaps the sidecar driver is
failing to notice that the sidecar kernel has removed a VP. This
causes a hang when onlining the CPU.

To make investigation easier, add some additional logging, plus poll
periodically in case the signal was dropped. This is not a final fix,
but it should make this condition easier to debug.

Signed-off-by: John Starks <jostarks@microsoft.com>